backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
neo4j==5.19.0
openai==1.30.0
python-dotenv==1.0.1
pydantic==2.7.1
numpy==1.26.4